backend/requirements.txt
absl-py==2.3.1
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
curl_cffi==0.11.4
cycler==0.12.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flatbuffers==25.2.10
fonttools==4.58.5
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
joblib==1.5.1
tensorflow==2.16.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.3.1  # ¡Cambiado a 0.3.1!
multitasking==0.0.11
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
scikit-learn==1.7.0
yfinance==0.2.65
peewee==3.18.1
pillow==11.3.0
platformdirs==4.3.8
protobuf>=3.20.3,<5.0.0
pycparser==2.22
Pygments==2.19.2
python-decouple==3.8
PyJWT==2.9.0
pyparsing==3.2.3